requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
python-dateutil>=2.8.2
flask>=3.0.0
flask-cors>=4.0.0
